ipc: share some utility functions for ipc4 path
Ipc3 & Ipc4 will share some component operations.
