+++ src/include/sof/ipc/topology.h
@@ -184,4 +184,25 @@ int32_t ipc_comp_pipe_id(const struct ipc_comp_dev *icd);
 int ipc_comp_dai_config(struct ipc *ipc, struct ipc_config_dai *common_config,
 			void *spec_config);
 
+/**
+ * \brief verify component params
+ * @param dev Component dev
+ * @param flag Component setting flag
+ * @param params Ipc params
+ * @return verify status
+ */
+int comp_verify_params(struct comp_dev *dev, uint32_t flag,
+		       struct sof_ipc_stream_params *params);
+
+/**
+ * \brief connect component and buffer
+ * @param comp Component dev
+ * @param comp_core comp core id
+ * @param buffer Component buffer
+ * @param buffer_core buffer core id
+ * @param dir connection direction
+ * @return connection status
+ */
+int comp_buffer_connect(struct comp_dev *comp, uint32_t comp_core,
+			struct comp_buffer *buffer, uint32_t buffer_core, uint32_t dir);
 #endif
