+++ src/ipc/CMakeLists.txt
@@ -8,6 +8,7 @@ endif()
 
 add_local_sources(sof
 	ipc-common.c
+	ipc-helper.c
 )
 
 if (CONFIG_TRACE)
