+++ test/cmocka/src/audio/pcm_converter/CMakeLists.txt
@@ -9,6 +9,7 @@ if(CONFIG_FORMAT_FLOAT)
 		${PROJECT_SOURCE_DIR}/src/ipc/ipc3/helper.c
 		${PROJECT_SOURCE_DIR}/test/cmocka/src/notifier_mocks.c
 		${PROJECT_SOURCE_DIR}/src/ipc/ipc-common.c
+		${PROJECT_SOURCE_DIR}/src/ipc/ipc-helper.c
 		${PROJECT_SOURCE_DIR}/src/audio/buffer.c
 		${PROJECT_SOURCE_DIR}/src/audio/pipeline/pipeline-graph.c
 		${PROJECT_SOURCE_DIR}/src/audio/pipeline/pipeline-params.c
